#369 [refactor] api 분리
